Use integration test because there is no push queue in unit tests
